# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImageOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          }
        },
        "required": [
          "id",
          "object_key",
          "entity_type",
          "entity_id",
          "expires_at",
          "created_at"
        ],
        "title": "ImageOut",
        "type": "object"
      },
...
      "ReadUrlRequest": {
        "properties": {
          "claim_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Token"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          }
        },
        "required": [
          "object_key"
        ],
        "title": "ReadUrlRequest",
        "type": "object"
      },
      "ReadUrlResponse": {
        "properties": {
          "read_url": {
            "title": "Read Url",
            "type": "string"
          }
        },
        "required": [
          "read_url"
        ],
        "title": "ReadUrlResponse",
        "type": "object"
      },
...
      "ReformUploadCompleteRequest": {
        "properties": {
          "claim_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Token"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "size_bytes": {
            "exclusiveMinimum": 0.0,
            "maximum": 10485760.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "object_key",
          "size_bytes"
        ],
        "title": "ReformUploadCompleteRequest",
        "type": "object"
      },
      "ReformUploadUrlRequest": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "size_bytes": {
            "exclusiveMinimum": 0.0,
            "maximum": 10485760.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "filename",
          "content_type",
          "size_bytes"
        ],
        "title": "ReformUploadUrlRequest",
        "type": "object"
      },
      "ReformUploadUrlResponse": {
        "properties": {
          "claim_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Token"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "required_headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Required Headers",
            "type": "object"
          },
          "upload_required": {
            "title": "Upload Required",
            "type": "boolean"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "object_key",
          "upload_url",
          "required_headers",
          "claim_token",
          "expires_at",
          "upload_required"
        ],
        "title": "ReformUploadUrlResponse",
        "type": "object"
      },
...
      "UploadUrlRequest": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "kind": {
            "enum": [
              "repair_shipping_upload",
              "custom_order",
              "sample_order",
              "quote_request"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "filename",
          "content_type"
        ],
        "title": "UploadUrlRequest",
        "type": "object"
      },
      "UploadUrlResponse": {
        "properties": {
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "object_key",
          "upload_url"
        ],
        "title": "UploadUrlResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "essesion api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/images/read-url": {
      "post": {
        "operationId": "create_read_url",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadUrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUrlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Read Url",
        "tags": [
          "images"
        ]
      }
    },
    "/images/reform-upload-url": {
      "post": {
        "operationId": "create_reform_upload_url",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReformUploadUrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReformUploadUrlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Reform Upload Url",
        "tags": [
          "images"
        ]
      }
    },
    "/images/reform-uploads": {
      "post": {
        "operationId": "register_reform_upload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReformUploadCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register Reform Upload",
        "tags": [
          "images"
        ]
      }
    },
...
    "/images/upload-url": {
      "post": {
        "operationId": "create_upload_url",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadUrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadUrlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Upload Url",
        "tags": [
          "images"
        ]
      }
    },
...
  }
}